requirements.txt
Django==5.1.4
djangorestframework==3.15.2
psycopg2==2.9.10
python-dotenv==1.0.1
isort==5.13.2